[front] fix: link to web extension on rate later page adapts to current browser
