+++ frontend/src/utils/extension.ts
@@ -1,0 +1,17 @@
+const getWebBrowser = () => {
+  if (navigator.userAgent.includes('Chrome/')) return 'chrome';
+  if (navigator.userAgent.includes('Chromium/')) return 'chromium';
+  if (navigator.userAgent.includes('Firefox/')) return 'firefox';
+  return 'other';
+};
+
+export const getWebExtensionUrl = (browser?: string) => {
+  const browserFamily = browser ?? getWebBrowser();
+  if (browserFamily === 'other') {
+    return undefined;
+  }
+  if (browserFamily === 'firefox') {
+    return 'https://addons.mozilla.org/en-US/firefox/addon/tournesol-extension/';
+  }
+  return 'https://chrome.google.com/webstore/detail/tournesol-extension/nidimbejmadpggdgooppinedbggeacla';
+};
